LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flex-item-compressible-002-expected.txt

Test1: "height: 100%"






Test2: "height: calc(100%)"






Test3: "height: calc(140px + 100%)"







PASS .flexbox 1
FAIL .flexbox 2 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="range" class="test1" data-expected-height="100">
    </div>
height expected 100 but got 129
PASS .flexbox 3
PASS .flexbox 4
PASS .flexbox 5
PASS .flexbox 6
FAIL .flexbox 7 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="range" class="test2" data-expected-height="100">
    </div>
height expected 100 but got 129
PASS .flexbox 8
PASS .flexbox 9
PASS .flexbox 10
FAIL .flexbox 11 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="text" class="test3" data-expected-height="140">
    </div>
height expected 140 but got 100
FAIL .flexbox 12 assert_equals:
<div class="flexbox">
      <div class="spacer"></div>
      <input type="range" class="test3" data-expected-height="140">
    </div>
height expected 140 but got 129
PASS .flexbox 13
PASS .flexbox 14
PASS .flexbox 15
